requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.15
django-admin-interface==0.19.2
django-appconf==1.0.5
django-ckeditor==6.5.1
django-colorfield==0.7.2
django-crispy-forms==1.14.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.8.0
django-js-asset==2.0.0
django-more-admin-filters==1.3
django-select2==7.10.1
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
psycopg2==2.9.4
pytz==2022.2.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
tzdata==2022.1
xlrd==2.0.1
xlwt==1.3.0